Add support for auto-detecting external Keystone endpoints in k8s
This adds new functionality to the Keystone container to allow it to
auto-detect external (public and admin) URLs when creating catalog
entries. This only works in Kubernetes environments with
KUBERNETES_RESOLVE_PUBLIC_ENDPOINTS=true and
KEYSTONE_SERVICE_NAME set to the name of the keystone service. The
end result of this is that Keystone can be used from outside of a
Kubernetes without any routing issues.

Services in preload.yml can also use the new functionality. By adding
`resolve: true` to an 'interfaces:' entry, the service will be looked
up in the Kubernetes API based on the URL and replaced with the
correct publicly-accessible node or ingress IP and port.