Add upgrade support to mysql-init
This adds support for basic schema upgrades in mysql-init. When an
existing schema is detected, mysql-init will now run (in order) all
schemas in `/mysql-upgrade.d/{major}.{minor}.{patch}/*.sql{,.j2}` when
when each version component is greater than the last-installed version
but less than or equal to the SCHEMA_XYZ_REV variables defined in the
Dockerfile.

This means that to relese a new schema:
 * The 'clean slate' /mysql-init.d/ should be updated to include any
   necessary changes
 * A new version directory should be added in `/mysql-upgrade.d/x.y.z`.
   This directory should contain differential changes needed to update
   the previous version to this new version.

Note that upgrade `.sql` scripts can also use j2 formatting when named
`.sql.j2`.