Enable cadvisor plugin by default in docker-compose
Currently the agent does not collect any metrics when running in
docker-compose. The docker plugin can't run now that the collector
is running as mon-agent (even with the socket mounted) and probably
needs a lot of refactoring to make it work well. As an alternative,
we now include cAdvisor by default and enable the cAdvisor plugin.

While the cAdvisor plugin only collects host-level metrics and does
not collect container metrics, it at least gets metrics flowing
through the system. Additionally, we can use cAdvisor as a base for
docker container monitoring in the future when the docker plugin is
refactored.