Add keystone-init container
This adds a new container capable of creating accounts in keystone.
It randomly generates passwords at first start and stores them in
Kubernetes secrets, bypassing Helm so that the secrets can be stored
securely.

Note that the container currently requires a Kubernetes environment,
support for running in docker will be added in the future. Eventually
this should replace most of our keystone bootstrapping logic.