[aggregator] Move config files to subdirectories
Due to a recent Kubernetes CVE, subpath mounting (which allowed
single-file mounts to `/`) is broken. As the old aggregator image
expected to find its config files in `/`, it won't run in affected
versions.

This adds a workaround to `start.sh` so `config.yaml.j2` is copied
from `/config` and `aggregation-specifications.yaml` is copied from
in `/specs`. This allows ConfigMaps to be mounted to these individual
directories without any trouble.

As this is a breaking change to the aggregator image, this is a new
minor version release.

Signed-off-by: Tim Buckley <timothy.jas.buckley@hpe.com>